Refectored tests for Node Class into seperate file NodeTests.cs
